refactor(queries-preview): move stub methods to a factory
EMP-1006
