+++ packages/x-components/src/__stubs__/index.ts
@@ -7,6 +7,7 @@ export * from './next-queries-stubs.factory';
 export * from './partials-results-stubs.factory';
 export * from './popular-searches-stubs.factory';
 export * from './promoteds-stubs.factory';
+export * from './queries-preview-stubs.factory';
 export * from './query-suggestions-stubs.factory';
 export * from './redirections-stubs.factory';
 export * from './related-tags-stubs.factory';
