+++ packages/x-components/src/x-modules/queries-preview/composables/__tests__/use-queries-preview.composable.spec.ts
@@ -1,31 +1,15 @@
 import { createLocalVue } from '@vue/test-utils';
 import Vuex, { Store } from 'vuex';
 import { DeepPartial } from '@empathyco/x-utils';
-import { QueryPreviewItem } from '../../store/types';
 import { installNewXPlugin } from '../../../../__tests__/utils';
-import { getQueryPreviewRequest } from '../../store/__tests__/utils';
 import { useQueriesPreview } from '../use-queries-preview.composable';
-import { getResultsStub } from '../../../../__stubs__';
 import { XPlugin } from '../../../../plugins';
 import { RootXStoreState } from '../../../../store';
 import { resetXQueriesPreviewStateWith } from '../../components/__tests__/utils';
 import { queriesPreviewXModule } from '../../x-module';
+import { createQueryPreviewItem } from '../../../../__stubs__/queries-preview-stubs.factory';
 
 describe('queries preview composables', () => {
-  const createQueryPreviewItem: (
-    query: string,
-    queryPreviewItem?: Partial<QueryPreviewItem>
-  ) => QueryPreviewItem = (query, queryPreviewItem) => {
-    const results = getResultsStub();
-    return {
-      results: results,
-      totalResults: results.length,
-      status: 'success',
-      request: getQueryPreviewRequest(query),
-      ...queryPreviewItem
-    };
-  };
-
   const localVue = createLocalVue();
   localVue.use(Vuex);
