+++ packages/x-components/src/x-modules/queries-preview/store/__tests__/utils.ts
@@ -1,6 +1,5 @@
 import { DeepPartial } from '@empathyco/x-utils';
 import { Store } from 'vuex';
-import { SearchRequest } from '@empathyco/x-types';
 import { resetStoreModuleState } from '../../../../__tests__/utils';
 import { queriesPreviewXStoreModule } from '../module';
 import { QueriesPreviewState } from '../types';
@@ -19,22 +18,4 @@ export function resetQueriesPreviewStateWith(
   state?: DeepPartial<QueriesPreviewState>
 ): void {
   resetStoreModuleState<QueriesPreviewState>(store, queriesPreviewXStoreModule.state(), state);
-}
-
-/**
- * Creates a queries preview search request stub.
- *
- * @param query - The query to set in the request.
- * @returns A {@link SearchRequest} to use with queries preview.
- *
- * @internal
- */
-export function getQueryPreviewRequest(query: string): SearchRequest {
-  return {
-    query,
-    rows: 3,
-    extraParams: {
-      extraParam: 'extra param'
-    }
-  };
 }
